refactor: move types to global
