+++ src/pages/main-page/main-page.tsx
@@ -6,7 +6,7 @@ import Map from '../../components/map/map';
 import OffersList from '../../components/offer/offers-list/offers-list';
 import UserMenu from '../../components/user-menu/user-menu';
 import { CardType, CityName } from '../../const';
-import { Offer } from '../../mocks/types/offers';
+import { Offer } from '../../types/offers';
 
 type MainPageProps = {
   offers: Offer[];
