+++ src/types/offers.ts
